feat: add homepage feature columns
